requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
chardet==3.0.4
idna==2.10
multidict==4.7.6
python-dotenv==0.14.0
selenium==3.141.0
slackclient==2.7.2
soupsieve==2.0.1
urllib3==1.25.9
yarl==1.4.2
